Fix 2013/mills + format/typo checks
The problem that was showing up is that one could connect and it would
show the time once but then it would close the connection and be unable
to connect again, showing an error in the browser. After looking at the
code it became clear it is because the backlog to listen(2) was 1 and
this is not enough for close(2) and then immediately listening again.

I updated the bugs.md file too - added a feature, not a bug. The author
notes that one cannot start the program right after termination and have
the program successfully bind the socket. They did note the fix and it's
well known (at least to those of us who are experienced in socket
programming :-) ) but as it was documented it is a feature not a bug. It
just means one has to wait for a few seconds (though 'few seconds' can
vary).

The README.md file has been format fixed.