Add 2014/wiedijk/try.sh + formatting/typo checks
The try.sh allows one to override the default indent(1) and sed(1) tools
and also verifies that they are executable. This means one can use, for
example, bcpp, instead of indent, or if one wants to specify a different
path to sed(1) they can. The README.md file explains this more.

README.md file format/typo checked.

This should complete the review of this entry.