Fix 2014/sinon + add try.sh + formatting
The mode where the game would play by itself did not work because the
code had 'sinon.c' in it but this does not exist. Now the code refers to
'prog.c'.

Add try.sh script that is an improvement over the previous try command
(or commands in a sequence) in that it loops, asking if you wish to try
again in case of a jam (or any other reason).

Format/typo check README.md.

This should complete the review of 2014/sinon.