Update tmp/things-todo.md
I noticed some issues in 2020/ferguson1/.winner.json that although is
for another issue I want to be sure to remember that this is one of the
ones that does in fact have a problem. Later on I can go through the
rest of them (not just mine but every one) and then make sure that
everything is okay.