Add 2000/dlowe/try.sh, update bugs.md, format
The try.sh script also has some fixes in the author's remarks (as in
they were fixed in the author's remarks) - it had "s in the command line
rather than single quotes but some of these command lines had * which
would expand to the files in the directory, causing errors.

The bugs.md has been updated with the issue of not printing 'poot' when
it should. I did some debugging and removed some (obviously nonsense)
ideas I had put in before (without thinking, being too tired, whatever
other reason or reasons). The real issue is a very strange one to be
sure.

I am pretty sure the reason it does not work with macOS (this was
already noted but worth repeating) is the version of perl as the author
noted this as an issue. We still have it as a known bug (some platforms)
but I don't think it's possible to fix without a higher perl version and
neither MacPorts nor Homebrew seem to have this.