Improve redefining keys in 1998/banks
The Makefile now has variables so one can do something like:

    make clobber IT=key DT=key ... all

The alt code has hard coded alternatives to page up and page down keys.
The reason it's hard coded is it would necessarily overcomplicate the
build process, requiring different macro names for alt and original
code, and since you can redefine them all in the original entry this is
not a problem. It is merely a convenience for those who are fine with
defaults. The replacement keys are both on the left side as it would be
harder to have to use left and right when trying to use the right side
for other keys (at least on a QWERTY keyboard?).

I tried to clean up the build instructions in the README.md file for
this too but this was a bit complicated to do. It might need a bit more
work.