Update todo file
There are still a few entries that need to be improved for the easier
redefining of macros (usually for alt versions) via the Makefile but as
this is tedious I have grown tired of it for the day. The thanks file
for some of these might need to be updated as well. The todo file
includes the entries modified today that have to be checked so this is
not lost though not many have to be updated in this way. This is
primarily for entries that this improves the build of the original entry
too like 1998/banks.