Fix major merge conflicts in thanks file
Unfortunately the renaming of the thanks-for-fixes.md to
thanks-for-help.md along with other changes with the it caused massive
merge conflicts. The file was out of order, there were a lot of >>>
lines etc.

I had to checkout the old file and update it for the changes I made this
morning. Unfortunately this does mean that the <a ..> anchors are no
longer there, relying on markdown to do that instead. That will have to
be fixed but I also noticed an error in that too. I am not sure if
that's because of the merge conflict or not though.

I will have to check that I got everything in order but I think I did as
I also added what I had added to it today.