Further fixes to thanks file
There was a line that was duplicated and more importantly than that some
entries that had been added yesterday before the merge conflicts were
resolved.

I believe this should complete the merge conflicts resolution and I
believe I got the '/' in front of the paths in all the right places. The
anchors (as in html <a ..>) should be in place as well.