Add 2015/burton/try.sh + format/typo checks
Renamed road.to.obscurity road.to.obscurity.md and make it markdown
format.

Add try.sh and try.this.txt which the script uses.

Format check README.md file, road.to.obscurity.md, spoilers.md.

The man pages have not been typo checked yet. The try.sh script might be
improved and there is another issue that is pending that prevents
completion of this entry (or potentially does) but it is mostly
complete.