Improve how to reconfigure coords in 2014/deak
The Makefile now has variables so that one can do something like

    make X1=-1 Y2=-1 clobber alt

or whatever, with the variables X1, X2, Y1 and Y2 set to what the author
originally had it as.

The alt code still has #ifndef..#endif so that one can (if they want to
be silly) compile it without make(1).