Add 2018/ciura/try.sh, missing PDF + formatting
The try.sh script does not use less for longer output because for some
reason less options -r/-R do not seem to work with the output (I haven't
looked into it in detail of how it works).

The missing PDF file was a paper I believe by the actual author but was
a dead link. Retrieved from the Internet Wayback Machine.

Format/typo check README.md file.

I believe this completes this entry's review.