improve FAQ 5.6, thank Cody for finding broken links
The question for FAQ 5.6 was improved to be more compact
while addressing both broken and wrong web links.

Removed any direct reference to drugs in an effort to
avoid problems from those who are humor impaired,
as well as "bots" to might flag the site as connected to drugs.

Moved the reference to the tool used by Cory to find
broken links to the thanks page.