Add 2019/yang/try.sh, update bugs.md + formatting
The try.sh script checks that the terminal size is large enough for two
files, the prog.c (last check) and sample_input.txt (provided by the
author). If it's too small for either it exits as it can't work. The
added ioccc.txt is very short in column and rows so that should not be a
problem (except in very extreme cases?).

The sample_input.txt file had trailing newlines removed because it
showed differences in output otherwise (and shouldn't).

The bugs.md was updated - add INABIAF for this entry.

Format/typo check README.md file.