Improve instructions/alt code for 1988/spinellis
There is no need to compile to 'runme': the way I designed the alt code
is so that it can be compiled and used exactly like the original entry.
Since the original entry need not compile to another file neither does
the alt version. It just compiles to spinellis.alt instead. (Perhaps
both should compile to something else though?)

The instructions of how to build/how to use in the README.md file have
been fixed and improved (it referred to the old way). The alt code
subsections have been added as well.

I also added to spinellis.alt.c #include <stdlib.h> as it's not a
one-liner anyway and this way system(3) is available without worrying
about warnings. The -Wno-everything still remains there in the system(3)
call in case there is a problem with the code that is typed in (though
that's not been tested either way).