Force use of -O0 -g in 2019/endoh
Previously only -g was after ${CFLAGS} but it needs to have -O0 in case
someone decides to run something like

    make clobber OPT= all

Use of OPT+= in the Makefile will not fix that problem. Thus if someone
does NOT mess with OPT or any other variables that are added to CFLAGS
it will look like:

     -O0 -g -O0 -g prog.c -o prog

Previously it was just '-g' and no '-O0' but if someone were to do
instead

    make clobber OPT=-O3 all

it would have messed that up.