Add 1989/ovdluhe/try.sh and fix Makefile
The try.sh script runs the program on the source file, the README.md and
the Makefile in a loop that runs four times. It has one loop so that
there is a better chance that it won't show the same output on the same
file more than once though it's certainly possible that it will.

The optimiser had been left disabled by accident and this has been
restored.