Add 1992/gson/try.sh, add note about optimiser
The optimiser must be disabled due to it otherwise crashing in macOS in
some invocations. This has been added to the Makefile, the use of -O0 is
forced and a note has been added to the README.md file.

The bugs.md file has been updated on details of how it uses gets() and
that text has been removed from the thanks file.

Minor format fixes in README.md.

The script added makes use of the author's script (that I added as
mkdict.sh for easier use) after first trying to find a dictionary file.
But one can override the dictionary file with the DICT environmental
variable.

This entry should now truly be done.