Re-enable optimiser in 2001/williams
The bugs.md was updated to remind people when debugging crashes that it
would be wise to have -O0 and -g (or if they can -ggdb3). That should be
noted in general for debugging but it's especially useful for crashes
and I'm trying to get this part done so I can resume 2019.