Improve 2019/karns/try.sh, force -O0
The try.sh script now prompts for input rather than just sleep for 2
seconds.

The README.md file has been format/typo checked.

Added issues reported by the author to bugs.md.

Force disable optimiser as the program doesn't work in some cases with
it enabled. If I recall correctly it crashed in some cases, something
the author noted happens sometimes (it is unknown if it is because they
had it enabled or if it was something else but I have not experienced a
crash since then).