Add 2019/lynn/try.sh, other files, updates
Add 2018/vokes/example-1.txt and 2018/vokes/example-2.txt as the author
refers to the files (and the try.sh uses them in the same way). This is
important so that the entry can be by itself (which is being
considered).

The README.md file has been format/typo checked.

The bugs.md has been updated but it refers to two sections in the
README.md file as there is a lot to include.