Add 2019/poikola alt code, try, try.alt, others
The alt code is so one can more easily filter the long output: it prints
a newline after each number is printed. The try.alt.sh script uses this.

Force -O0: the author stated that with clang they got correct output for
[0123s] but with GCC 6 they only got that with level 0. Since 0 works
with clang and since the author stated that the compilers used were
picky I changed it to level 0. Similarly and for the same reason I
changed the C standard to gnu11 from gnu17.

Format/typo check README.md file.

Updated bugs.md: noted that the max file size is 1GB.

I would like to give two quotes from the person this entry was dedicated
to:

    'When you are about to jump, you are all alone and have to make your own
     decisions. Up there, it's all "up yours".'

and

    'I will move to Copenhagen and apply for Swedish citizenship.'

simply because they're funny and ironic (and one is quoted by the
author). These might be found somewhere in this commit.