Improve 2020/ferguson2 try scripts
Added key.txt which is based on the judges' try.this.txt file (which was
based on something I originally submitted) which only has the key rather
than the input. This file is then used by try.sh and try.alt.sh.

Pass -v to recode now so that one can see the settings used and in a
nice way.