Add 1986/stein/stein.sh
This script runs the two commands suggested in order to get it to output
cleanly. It is suggested other systems might need another set of
characters but one can also get away with just using ./stein so this is
also done but at the end of the script to not cause any problems (it
doesn't show a newline at output).