Fix 2020/ferguson1/Makefile
The DATA variable still had html files which I removed a while back.
This made make fail.

Also there is a joke Easter egg that I added: something to do with
eating as that's also kind of the point of the entry: though eating
insects (or bugs for a joke I made originally, so that it can say there
are bugs that are not bugs). Snakes are carnivores: they don't eat
apples and although insects are not going to be filling enough they can
eat them. But what the joke is one has to discover for themselves!