Convert to/format markdown in 1992/buzzard.2
The file buzzard.2.README.md was formatted in markdown style.

The file buzzard.2.design was turned into a markdown file,
buzzard.2.design.md, and formatted in markdown style. This one was
problematic in some ways but I think it's sorted out now. In a few
places it was not entirely clear if certain things should be in code
blocks or not but they were made to be as they were on the same line as
code (after). It doesn't appear to be a comment in the syntax of
comments that later appears, however. Another one I was not sure about
is if the author meant '{ .. .. }' to be a way to say this is code
without the {}s or if that should be part of the code. As these are not
clear there might be some mistakes.