Improve message in all rule in 1993/lmfjyh
It now clarifies that in the likely case it fails to compile the alt
version will be built as lmfjyh itself rather than suggesting one run
make alt (which basically did already). It does show how one can compile
the alt code manually still but it is not strictly necessary as this
entry is one of the rare cases where if the primary target fails the alt
target is built automatically.