Release new ioccc_status.sh
New version: 0.0.3-0 2024-01-15.

The following options were added:

    -N		set dry run mode: just show what would be changed (def: make changes)
    -S sed	set path to sed(1) (def: $SED)
    -G grep	set path to grep(1) (def: $GREP)

where $SED and $GREP are set previously by

    SED="$(type -P sed)"
    GREP="$(type -P grep)"

I am not happy with -N: I'd like it to be -n to match other tools.
However the -n is already used to update the news date. Since doing that
is far more common (in use) than a dry-run I have the dry-run as a
majuscule ('upper case') letter and the news update date option as a
minuscule ('lower case') letter.

The verbosity level check was slightly modified too. If >= 1 (was > 1)
then show what will change (assuming any change requested). If -N
specified then it shows what WOULD HAVE changed (and then exits). If
verbosity > 1 it shows what (for those with a requested change) was
changed (and to what).

It is known that the status.json might have a different format so this
commit might be premature (and I might need to rewrite the tool
entirely) but it was on my mind so I decided to do it as I am not awake
enough to do other things.