Remove \ from $ in code blocks in 2020/carlini
I believe it was a mistake to add the \ before the $ as it was in a code
block. In one case a pair of '`'s was missing and those were added there
so that both are now code blocks. If it is not an error to have the '\'
before the '$' then the code block added should still be there but I
believe it was a mistake anyway.