Add 1995/savastio/try.sh and update bugs.md
The try.sh script demonstrates both positive numbers and input that is
not a number at all (comparing two different strings).

The bugs.md has a new INABIAF - negative numbers will hang the program.
It is not clear if another version should be added to gracefully handle
this but that can be decided another time. It is not a documented
feature so it might be called a bug to fix.