Improve 1995/garry try scripts
They now show the garry.data file and also show a bit more of what is
being done (in the script).

In the thanks file I removed the details as to why there are two
scripts, one for the main version and one for the alt, as that was only
there as at one point using environmental variables was the way it was
being done, but that is no longer important or useful to know.