Make usage of 1998/dloweneil more inclusive
This is a cosmetic change only and rather unimportant but I have to be
complete: when I added the alt code I used the word anticlockwise (as I
would) instead of counterclockwise. But since the original entry
referred only to counterclockwise I decided to make it use both like the
alt code does.