Add 2000/natori/try.alt.sh + improve try.sh
The try.alt.sh script is slightly different from try.sh in that the
order is reversed.

The try.sh script now at the end always starts out in a sequence of
northern hemisphere then southern hemisphere so that one can guarantee
to get that form. The try.alt.sh does the same but starts out with the
southern hemisphere first.