Add 2006/monge/try scripts
The alt script will reconfigure the width, height and iterations for the
second invocation.

Note that in some platforms this entry will not work (this is
documented in the remarks). That is one of the reasons there is only one
expression tested as one of the platforms it will not work with is arm64
which is what I primarily use.