Add 2006/stewart/try.sh
It will run the program on a few files but before it runs it on the
remaining files (as there are quite a lot) it will ask you if you wish
to do so. It only asks once. If convert is found it will run convert on
the xbm images to convert to png. If it's not installed it will only
warn the user once so as to not spam them. If convert fails (say because
the convert is not convert from ImageMagick) it will warn once at the
end of the script. In any event if any files were converted it will list
the png files that exist at the end of the script. Otherwise it'll list
the xbm files. It does not first check if any of those file types exist
as it is presumed that at least a few will exist. If none do it is
likely an error in the program or a corrupted data file.