Add 2006/borsanyi/try.sh
The script runs in a loop until the user answers anything but 'y' or 'Y'
when asked if they wish to make another one. It does NOT check the valid
range of characters or the length. Instead it suggests one try and see
what happens if they do not follow the rules.

Updated .gitignore.