Improve 1994/westley alt + to use etc.
The to build/original build, to use/original use and try/original try
sections have been changed in the README.md to be much more consistent
and clearer (and detailed in some ways).

The to run section (not the original run section as that one is not
supposed to be compilable and is not compilable either) has
'./westley.alt' now as that will give you a message even right after
running 'make clobber alt'.

The alt code was changed in that blank lines were removed and (as a
cosmetic change) the 'return 0;'s were changed to 'exit(0);'s.
Originally the idea was to have 'int main()' be 'main()' (like the
original in the #if 0..#else..#endif block) but this will
trigger -Wimplicit-int and since one would not want to run a more
complicated 'make' invocation (instead of a simple 'cc' invocation or at
least a simpler one) it means the output is messed up with a warning.
But using exit(0); is fine albeit with an additional #include.

The try.alt.sh script now has an -o westley.alt and instead of running
./a.out it runs ./westley.alt. This makes it more like normal 'make alt'
targets.

The Makefile all target now has information about the alt build as it
does not have the restrictions of column width, character tab length and
wraparound.

The .gitignore no longer has a.out in it.