Improve and fix 2011/fredriksson/try.sh
The script now writes the output to a file (via tee(1)) and uses diff -s
on the file and fredriksson.c. It also fixes a problem where if one were
to hit space after the prompt (lines would be merged) rather than
enter/return.

Fixed shellcheck warning (no need to ignore it).