Further improvements to 2012/blakely/try.sh
The script now has quotes surrounding commands with '*'s in it and the
messages (of which some were added - only one existed before) were
improved when it comes to giving a warning that the next command might
take a bit of time. There are different messages depending on how slow
the commands are though this obviously cannot be even close to perfect.

Updated .gitignore.