Improve 2012/deckmyn/try.sh
Improve less(1) options so that the screen is not cleared (to keep the
source visible after it's seen) and to exit at EOF.

Try converting image files to png files. If this fails then as long as
pbm files exist it will print those (telling the user to open them in
their graphics viewer of choice). Otherwise if png files exist it lists
the png files (with the same message). If no png and no pbm files exist
it warns them that something went wrong and they are free to fix and
open a pull request. This handling of convert(1) checks should probably
be done in other scripts where I neglected to do this (it's entirely
possible though unlikely that a convert program exists that returns
success but does not do what we expect).

Updated .gitignore.