Improvements to build and script of 2013/endoh4
The Makefile was modified in such a way that allows one to easily
redefine the size (in the code the S macro). For instance:

    make clobber SIZE=50,50 all

The try.sh was renamed to endoh4.sh as it's more of a wrapper to the
program rather than things to try. Also it now lets one specify more
than one file to feed to the program (in sequence).

The bugs.md was updated - added INABIAF. If the file format is invalid
it will very likely crash.