Improve 2014/morgan/try.sh
After copying prog to morgan (which was already done) it removes prog
and then uses morgan to build prog. Following this it shows that they're
identical to each other. The rest of the script it uses morgan, not
prog, and it's the same as before (what it runs) except that the use of
'ls' is no longer there as there is no need for it.