Fix and improve 2014/sinon scripts
The try script now runs the demo mode if perl is installed. It also now
prompts before running each script (the below noted ones) so that one
can control the pace of the output a bit more.

The glock.sh and hecate.sh scripts could only be run once successfully
because the output of run.c overwrites it (with non code). First one has
to do cp -f prog.c run.c which the scripts now do. This was not a
problem in the try script because I added that step explicitly (and is
now not needed or there).