Update Ferguson chocolate cake recipe
More notes have been made bold and capitalised as well as more explicit
and clearer.

Removed things that are not relevant (directly or indirectly) to the
recipe itself. For instance although fun the other IOCCC etc. just makes
the text longer and not get to the recipe which is the good part and
which was my original intent (obviously). For reference here that refers
to:

    International Office of Cocoa, Chocolate and Sugar Confectionery

which is 'the other' IOCCC. A PDF file that was related to that was
removed which I might put elsewhere in time (other things removed might
or might not be put elsewhere too but that's none of that is certain
except I have copies of everything just in case I do wish to do that).

This commit will require a manifest update but that will likely not
happen until Saturday.