Further refinements to Ferguson chocolate cake
Be even more clear about certain notes that are vital to success in
making the cake. Of course in 2020/ferguson2 it's enciphered with the
Enigma machine (was part of a fun way to explore the entry but obviously
very few would do that since it's in plaintext in 2020/ferguson1).

The file cake.jpg was renamed to ioccc-cake.jpg. This file has a feeble
attempt to draw IOCCC on it but it does not show the cake iced properly
so I feel like it shouldn't be the main image. This image is linked to
but it is not embedded in the document.

The new cake.jpg is one from 14 February (my birthday) in 2022 (2022
because it's the best picture I found in recent years that show both the
top and the side so one can see how thick/tall it is or better see it
anyway). This should improve the presentation of the cake and do it much
more justice.