Update todo file
Removed an item that I already finished.

Merge two items that were related.

Added an item (does that mean I didn't remove an item?) for
2018/ferguson that has to be done after the other tasks are done and on
specific systems (was done for personal reasons at submission but since
the judges referred to it in their remarks I have kept it and will
continue to keep it).