Add 2020/endoh3/run_clock.alt.sh for alt code
The alt code provided by the author back in 2020 uses Unicode but there
wasn't a script for it (the script run_clock.sh which I made some fixes
in last year was added by the judges but no version for the alt code was
added).

The scripts also have another fix in them: make clock or make clock.alt
first as if not and the program does not exist the script will happily
overwrite the C files and prevent it from being compiled. They use 'make
-B' just to be sure that it's compiled.